# docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: Layblar Platform API
  description: "# Introduction\nThis is the API Documentation for the Layblar-Platform.\
    \ \n"
  contact:
    name: Layblar Documentation
    url: https://layblar.github.io/docs
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
  x-logo:
    url: https://layblar.github.io/docs/img/layblarlogo.svg
servers:
- url: "http://{tenant}"
  variables:
    tenant:
      default: localhost:8080
...
paths:
  /auth/login:
    post:
      tags:
      - Authorization
      summary: Login
      description: Login with given credentials
      operationId: loginUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountLoginDTO'
      responses:
        "401":
          description: Unauthorized
        "403":
          description: Invalid User
        "500":
          description: Server Error
        "200":
          description: The logged in user
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
      security:
      - none: []
  /auth/register/householduser:
    post:
      tags:
      - Authorization
      summary: Register a new Household user
      description: Creates a new user and assigns them to a new Household
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterHouseholdUserDTO'
      responses:
        "401":
          description: Unauthorized
        "403":
          description: Invalid User
        "500":
          description: Server Error
        "201":
          description: The registered user
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                  researcherId:
                    type: string
                  email:
                    type: string
                  householdId:
                    type: string
                  roles:
                    type: array
                    items:
                      type: string
      security:
      - none: []
  /auth/register/researcher:
    post:
      tags:
      - Authorization
      summary: Register a new researcher
      description: Creates a new researcher
      operationId: createResearcher
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterResearcherDTO'
      responses:
        "401":
          description: Unauthorized
        "403":
          description: Invalid User
        "500":
          description: Server Error
        "201":
          description: The registered researcher
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                  researcherId:
                    type: string
                  email:
                    type: string
                  householdId:
                    type: string
                  roles:
                    type: array
                    items:
                      type: string
      security:
      - none: []
  /deviceLibrary:
    get:
      tags:
      - Device-Library
      summary: List devices
      description: Get a list of devices from the device library
      operationId: listDevices
      parameters:
      - name: name
        in: query
        schema:
          type: string
      responses:
        "401":
          description: Unauthorized
        "403":
          description: Invalid User
        "500":
          description: Server Error
        "200":
          description: List of devices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeviceDTO'
      security:
      - none: []
  /deviceLibrary/category:
    get:
      tags:
      - Device-Library
      summary: List all devices categories
      description: Get a list of device categories from the device library
      operationId: listCategoires
      parameters:
      - name: name
        in: query
        schema:
          type: string
      responses:
        "401":
          description: Unauthorized
        "403":
          description: Invalid User
        "500":
          description: Server Error
        "200":
          description: List of device categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeviceCategoryDTO'
      security:
      - none: []
  /deviceLibrary/category/{categoryId}:
    get:
      tags:
      - Device-Library
      summary: Device category by Id
      description: Get a Device category by a Id
      operationId: getCategoryById
      parameters:
      - name: categoryId
        in: path
        required: true
        schema:
          type: string
      responses:
        "401":
          description: Unauthorized
        "403":
          description: Invalid User
        "500":
          description: Server Error
        "404":
          description: Device category not found
        "200":
          description: Device category
          content:
            application/json:
              schema:
                type: object
                properties:
                  deviceId:
                    type: string
                  deviceName:
                    type: string
                  deviceDescription:
                    type: string
                  manufacturer:
                    type: string
                  modelNumber:
                    type: string
                  powerDraw:
                    format: int32
                    type: integer
                  energyEfficiencyRating:
                    type: string
                  weight:
                    format: float
                    type: number
                  categories:
                    type: array
                    items:
                      $ref: '#/components/schemas/DeviceCategoryDTO'
      security:
      - none: []
  /deviceLibrary/{deviceId}:
    get:
      tags:
      - Device-Library
      summary: Device by Id
      description: Get a device by a Id
      operationId: getDeviceById
      parameters:
      - name: deviceId
        in: path
        required: true
        schema:
          type: string
      responses:
        "401":
          description: Unauthorized
        "403":
          description: Invalid User
        "500":
          description: Server Error
        "404":
          description: Device not found
        "200":
          description: Device
          content:
            application/json:
              schema:
                type: object
                properties:
                  deviceId:
                    type: string
                  deviceName:
                    type: string
                  deviceDescription:
                    type: string
                  manufacturer:
                    type: string
                  modelNumber:
                    type: string
                  powerDraw:
                    format: int32
                    type: integer
                  energyEfficiencyRating:
                    type: string
                  weight:
                    format: float
                    type: number
                  categories:
                    type: array
                    items:
                      $ref: '#/components/schemas/DeviceCategoryDTO'
      security:
      - none: []
...
components:
  schemas:
    AccountLoginDTO:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
...
    DeviceCategoryDTO:
      type: object
      properties:
        deviceCategoryId:
          type: string
        deviceCategoryName:
          type: string
        deviceCategoryDescription:
          type: string
    DeviceDTO:
      type: object
      properties:
        deviceId:
          type: string
        deviceName:
          type: string
        deviceDescription:
          type: string
        manufacturer:
          type: string
        modelNumber:
          type: string
        powerDraw:
          format: int32
          type: integer
        energyEfficiencyRating:
          type: string
        weight:
          format: float
          type: number
        categories:
          type: array
          items:
            $ref: '#/components/schemas/DeviceCategoryDTO'
...
    RegisterHouseholdUserDTO:
      type: object
      properties:
        password:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
    RegisterResearcherDTO:
      type: object
      properties:
        password:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        organization:
          type: string
...
  securitySchemes:
    SecurityScheme:
      type: http
      description: Authentication
      scheme: bearer
      bearerFormat: JWT
    jwt:
      type: http
      description: Authentication
      scheme: bearer
      bearerFormat: JWT
    none:
      type: http
      description: No Authentication
      scheme: No-Auth