# gateway/src/main/java/at/fhv/layblar/authentication/dto/RegisterHouseholdUserDTO.java
package at.fhv.layblar.authentication.dto;

public class RegisterHouseholdUserDTO {

    public String password;
    public String email;
    public String firstName;
    public String lastName;
    
}